requirements.txt
transformers==4.45.2
torch==2.9.0
torchaudio==2.9.0
--extra-index-url https://download.pytorch.org/whl/cu126
streamlit==1.44.1
openai-whisper==20250625
huggingface-hub==0.36.0
accelerate==0.27.2